- separete agents cards from table - improve file naming - change routes to be leaderboard/agents - add metadata og image - update nav leaderboards link
